not sure why revert is not working

+++ controller.runmean.cc
@@ -210,7 +210,7 @@ void Controller::refineParameters(const uint64_t sequence_number_acked,
   }
   double mrtt=diffsum/((int)rtimes.size());
   fprintf(stderr,"rttmean: %i\n",(int)mrtt);
-  double bwest=runmean.size()/resolution;
+  double bwest=((double)runmean.size())/resolution;
   /*double slope = 0.5414;
   double icept = -1.0402;
   double tfbest = 2*sqrt(runmean.size()+3/8)*slope+icept;
